changes from code review
- bubbling-up
- moves eventQueueMaxSize to config
- removes reference to unimplemented ModifyFieldSelection
- adds comments for the global event-PDS list
