+++ examples/CMakeLists.txt
@@ -247,6 +247,5 @@ if(UA_ENABLE_PUBSUB)
     if(UA_ENABLE_PUBSUB_EVENTS)
         add_example(tutorial_pubsub_publish_event pubsub/tutorial_pubsub_publish_event.c)
         add_example(tutorial_pubsub_subscribe_events pubsub/tutorial_pubsub_subscribe_events.c)
-        add_example(tutorial_pubsub_subscribe_events_request_metadata pubsub/tutorial_pubsub_subscribe_events_request_metadata.c)
     endif()
 endif()
